fix: allow button to be clicked in React with IE11
When used in a React application in IE11, the button was not
responding to clicks. Adding a no-op click handler on the button within
the component fixed the issue and does not appear to have any negative
side effects in a non-React environment.

Tested on Polymer demo page and in OrionReactDemo in Firefox,
Chrome, and IE11.