feat: add new loading state
This change adds a new loading state which utilizes auro-loader.
Disabled state has changed how it applies opacity - rather than a general button opacity
the border, bg and font colors are converted to an rgba value with the same opacity in it.
This allows for the loader to not have an opacity even when the button is disabled.